% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoDev.R
\name{anoDev}
\alias{anoDev}
\title{Conduct an analysis-of-deviance for the MSU pollinator analysis.}
\usage{
anoDev(df, ft, annuals = TRUE, blocks = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} containing columns for \code{PollinatorVisits},
\code{Cultivar}, \code{SamplingPeriods}, and \code{FlowerType}.  Typically,
\code{Cultivar} is nested within \code{FlowerType}.}

\item{ft}{A \code{FlowerType} within the Perrenial dataset.}

\item{annuals}{A \code{logical} indicating if annuals or perennials are the
analysis type of focus.  Default is \code{TRUE}.}

\item{blocks}{A \code{logical} indicating if a random block effect shoud be
included.  Requires the presence of a \code{Block} variable.  Default is
\code{TRUE}.}
}
\value{
A \code{list} containing results of the original fit, adjusted
  Tukey-like pairwise multiple comparisons, and the analysis-of-deviance.
  The four constituent objects are \enumerate{

\item \code{a} The resulting \code{data.frame} housing the statistical
results of the analysis of deviance.  Recall that the degrees of freedom
reports numbers one less than the distinct groups tested.  The statistic
entitled \code{phi} is the (Pearson's chi-squared) overdispersion.  These
statistics are corrected for quasilikelihoo.

\item \code{c} The Tukey-lettering system identifying how individual levels
compare.  Utilized \code{multcomp:cld}.

\item \code{i} The inverted confidence intervals on the original scales of
the pairwise estimates.

\item \code{m} The original \code{glm} fit, assuming a Poisson distrubtion
and possible offset.

\item \code{t} The original mutliple-comparison \code{glht} fit, assuming
Tukey-like pairwise comparisons.

}
}
\description{
Conduct an analysis-of-deviance for the MSU pollinator analysis
  assuming a Poisson-distributed outcome and varying exposure.
}
\details{
Function \code{stats::glm} performs the analysis-of-deviance via use
  of the \code{quasipoisson} family, and thus \eqn{log} link. Following
  initial fit, a chi-squared test is used to test via deviance.  This
  generalized-linear-model set-up corresponds to the typical partial
  \eqn{F}-test conducted via residual error in the general-linear-model
  normal-outcome case.

  Note that the use of the \code{quasipoisson} family provides a conservative
  estimate of underlying test statistics.  This is due to the
  \code{quasipoisson} correcting for observed overdispersion in
  \code{poisson}-model fits.
}
\seealso{
\code{stats::glm}, \code{stats::anova}, \code{stats::confint},
  \code{multcomp::glht}
}
\author{
Jason Mitchell, based on original work by Zoe Gustafson.
}
