% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpFormula.R
\name{cleanUpFormula}
\alias{cleanUpFormula}
\title{Massage a provided vector of intended model character strings into the
  format necessary for a \code{formula}-class object.}
\usage{
cleanUpFormula(var)
}
\arguments{
\item{var}{The character vector of data intended to be used and formatted as
"clean" formula strings (as least in terms of the R-object
\code{formula}-class.)}
}
\value{
The original character vector \code{var} with slashes "\code{/}"
  replaced with a single underscore "\code{_}," hyphens "\code{-}" replaced
  with a double underscore "\code{__}," and commas "{,}" replaced with a
  triple underscore "\code{___}."
}
\description{
Massage a provided vector of intended model character strings
  into the format necessary for a \code{formula}-class object.
}
\details{
As coded here, the \code{out} object is assumed to possible contain
  a "hyphen," serving as a minus sign, as well as commas, due to typical
  syntax involving fits against the \code{binomial} family and associated
  \code{cbind} usage.  It is possible that the syntax of replacement steps
  performed here is not ideal for your situation.
}
\references{
Wilkinson G.N., and Rogers C.E. (1973). "Symbolic Description of
  Factorial Models for Analysis of Variance." Applied Statistics, 22, 392-399
}
\author{
WEST, Inc.
}
