% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{fits}
\alias{fits}
\title{Fit all options for a given outcome model scenario, allowing
  fixed-effect environmental covariates to vary.}
\usage{
fits(mStrings, CN, CNMax = 30, row, df, offsetVar, saveModels = FALSE)
}
\arguments{
\item{mStrings}{A list housing various model \code{formula} strings from
function \code{makeModelStrings}.}

\item{CN}{A numeric vector of results, as output via \code{sapply} and
\code{getKappa} from the \code{buildModel} wrapper function.}

\item{CNMax}{A number specifiying the maximum condition number for which
models are to be evaluated.  Default is \code{30}.}

\item{row}{A one-row \code{data.frame}, typically originating from a
\code{csv} housing desired model characteristics.  Should include at least
\code{outcome}, \code{fitFunction}, \code{outcomeFormula},
\code{possibleCovars}, \code{baseFormula}, \code{ziFormula}, and
\code{family}.}

\item{df}{A dataframe containing the variables necessary to fit a Piepho &
Ogutu linear mixed model.}

\item{offsetVar}{The variable to be utilized as an offset, submitted as a
character string.  Typically \code{SamplingPeriods}.  Default is
\code{NULL}, or no offset.}

\item{saveModels}{A logical indicating if each individual model should be
pushed as functional output.  Default is \code{FALSE}.}
}
\value{
A list with several elements.

\describe{
\item{good}{A list of length equal to the number of models whose condition
number was less than the \code{CNMax}.  Entries are direct model objects from
fitting, and thus contain all information possibly of value.}
\item{bad}{A list of length equal to the number of models whose condition
number was greater than the \code{CNMax}.  Entries simply report the
offending condition-number values.}
\item{mess}{A list of length equal to the length in list \code{good}.
Entries correspond to messages recorded during model fits.  Values of
\code{NULL} mean no message was recorded.  Typically, this implies model
convergence.}
\item{fitsL}{A list of length equal to the number of distinct models included
via \code{mStrings$modelRaw} housing results from the model fit.  In the case
the condition number was greater than \code{CNMax}, the number is the
offending condition number. Otherwise, an object of the class peculiar to the
model-type fit is returned.  In the case that \code{saveModels} is
\code{FALSE}, a simple character string is returned instead.}
\item{aic}{A list of length equal to the length in list \code{good}. Reported
numbers are the AIC, as pulled from the \code{summary} of the original model
fits.  Entries with \code{9999999999} generally represent models for which
convergence was achieved, but for which the AIC was reported to be \code{NA}.
This means these models could be included within the \code{good} list. }
\item{top}{A \code{data.frame} encapsulating all other results in a simple
tabulated format.  Conducive for output to \code{csv} format.}
}
}
\description{
Fit all options for a given outcome model scenario, allowing
  fixed-effect environmental covariates to vary.
}
\details{
Function \code{fits} fits the model functional form specified via
  \code{row}.  Included fit models are dictated via \code{mStrings} with
  calculated condition numbers less than \code{CNMax}, whose default is set
  at \code{30}.

  Although function \code{buildModel} constructs model strings both with and
  without an intercept, formula manipulation to create the \code{top}
  dataframe currently assumes the presence of one.  Thus, manually removing
  the intercept via the typical \code{-1} will probably lead to an error.

  Currently, count models always assume an offset of \eqn{log(30)} (for 30
  m\eqn{^2} searched per site visit).
}
\seealso{
\code{makeModelStrings}, \code{getKappa}
}
\author{
Jason Mitchell
}
