% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKappa.R
\name{getKappa}
\alias{getKappa}
\title{Calculate the condition number of a set of covariates.}
\usage{
getKappa(x, df)
}
\arguments{
\item{x}{A character string containing covariates present in \code{df},
separated by \code{" + "}.}

\item{df}{A dataframe containing the data to be used for subsequent modeling.}
}
\value{
The condition number, a number within the range \eqn{1,\infty)}.
}
\description{
Calculate the condition number of a set of covariates.
}
\details{
Included covariates stored as class \code{factor} within \code{df}
  are expanded via function \code{model.matrix} via the use of reference-cell
  coding. Thus, a \eqn{k}-level factor contributes \eqn{k - 1} variables in
  the design matrix ultimately used in modeling.

  Currently, the function does not check if each variable within \code{x} is
  actually contained within \code{df}.
}
\references{
Belsley, D.A., Kuh, E., and Welsch, R. E. (1980).  Regression
Diagnostics: Identifying Influential Data and Sources of Collinearity.  New
York, NY: John Wiley \& Sons, Inc.
}
\author{
WEST, Inc.
}
