% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoissonMeans.R
\name{getPoissonMeans}
\alias{getPoissonMeans}
\title{Estimate \code{Cultivar} means via use of Poisson regression.}
\usage{
getPoissonMeans(df)
}
\arguments{
\item{df}{A \code{data.frame} containing columns for \code{PollinatorVisits},
\code{Cultivar}, \code{SamplingPeriods}, and \code{FlowerType}.}
}
\value{
A \code{data.frame} containing estimates of means and 95\% confidence
  intervals.  In the case a \code{Cultivar} is all-zero, the confidence
  interval is returned as \code{NA}.
}
\description{
Estimate \code{Cultimar} means via use of Poisson regression,
  with any overdispersion handled via quasilikelihood.
}
\details{
Function \code{getPoissonMeans} runs a simple intercept-only model
  for each \code{Cultivar} found within \code{df}.  It also incorporates
  offsets via \code{SamplingPeriods}.  This allows application of function
  \code{stats::confint} so that resulting 95\% confidence intervals are fit
  on the log-scale.  Practically, this ensures they do not cross zero.

  Because function \code{getPoissonMeans} is geared towards estimate Poisson
  means, no consideration of overdispersion is made.
}
\seealso{
\code{stats::glm}, \code{stats::coef}, \code{stats::confint}
}
\author{
Jason Mitchell, based on original work by Zoe Gustafson.
}
