% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lettered.R
\name{lettered}
\alias{lettered}
\title{Create a Tukey-type bar chart with Tukey letters.}
\usage{
lettered(
  df,
  here,
  fileSave,
  title,
  xlab,
  ylab,
  ft,
  out,
  year,
  height,
  width,
  letters = TRUE,
  bounds = FALSE,
  text = 11
)
}
\arguments{
\item{df}{A \code{data.frame} containing columns for \code{FlowerType},
\code{Variety}, \code{b0}, \code{lo2.5}, and \code{hi97.5},
\code{Captured}, and \code{Observed}.  Typically the output from function
\code{getPoissonMeans}.}

\item{here}{A text string indicating the subfolder into which results should
be saved.  Also sometimes used to help entitle output.}

\item{fileSave}{A character string used to name the plot file.}

\item{title}{A character string used to entitle the plot.}

\item{xlab}{A character string used to label the \eqn{x}-axis.}

\item{ylab}{A character string used to label the \eqn{y}-axis.}

\item{ft}{A text string indicating the \code{FlowerType} of focus.}

\item{out}{A character string used to identify the directory into which
output is to be saved.  The directory should have a trailing slash
"\code{/}."}

\item{year}{The year of focus.  Typically one of \code{2017} or \code{2018}.}

\item{height}{A number indicating the height of the resulting \code{png}, in
inches.}

\item{width}{A number indicating the width of the resulting \code{png}, in
inches.}

\item{letters}{Default is \code{TRUE}.}

\item{bounds}{Default is \code{FALSE}.}

\item{text}{The size of the output font.  Larger means bigger.  Default is
\code{11}.}
}
\value{
A plot with Tukey lettering, assuming \code{letters} is set to
  \code{TRUE}.
}
\description{
Create a Tukey-type bar chart with Tukey letters.
}
\details{
Function \code{lettered} ensures that "the letters" appear above
  bars representing means.  The letters indicate the groupings that are
  statistically similar. This plot emulates the example plot from Garbuzov &
  Ratnieks (2014).
}
\references{
Garbuzov and Ratnieks.  2014.
}
