% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModelStrings.R
\name{makeModelStrings}
\alias{makeModelStrings}
\title{Given a character vector of length \eqn{J} of variable names,
  construct all \eqn{K <= J} combinations of possible model covariate
  strings.}
\usage{
makeModelStrings(row, K, offset = NULL)
}
\arguments{
\item{row}{A \code{data.frame} of one row, containing several columns used to
build a \code{formula} object.  See Details.}

\item{K}{An integer less than or equal to the length of variables provided
via \code{possibleCovars}.  Currently, only values of \code{1}, \code{2},
\code{3}, \code{4}, or \code{5} are allowed.}

\item{offset}{A character vector of variables to include as an offset.
Default is \code{NULL}.  Not currently coded to do anything in the case of
a non-\code{NULL}.}
}
\value{
A list with three objects:

  \describe{
  \item{\code{modelIntY}}{A character vector housing \code{formula} objects for
  use in modeling with intercepts "{ + 1}" explicitly included.}
  \item{\code{modelIntN}}{A character vector housing \code{formula} objects for
  use in modeling without intercepts "{ - 1}" explicitly included.}
  \item{\code{modelZI}}{A character vector housing \code{formula} objects for
  use in modeling.  Presence or absence of intercepts dictated via underlying
  stored \code{csv} from which \code{row} originates.}
  \item{\code{modelRaw}}{A character vector housing \code{modelIntY} (or
  \code{modelIntN}) without an intecept summand.}
  }
}
\description{
Given a character vector of length \eqn{J} of variable names,
  construct all \eqn{K <= J} combinations of possible model covariate
  strings.
}
\details{
Function \code{makeModelStrings} simply uses the \code{combn}
  function to select all combinations of required model covariate strings.

  Required object \code{row} should include columns for each of

  \describe{
  \item{\code{possibleCovars}}{A character vector housing the names of the
  different variables for which univariate models have previously been fit.}
  \item{\code{outcomeFormula}}{A character vector housing the name(s) of the
  variable used as an outcome, and intended for use within a resulting
  \code{formula}-type object. Can be fancier "\code{cbind}-type" character
  strings in the case of binomial outcomes.}
  \item{\code{baseFormula}}{A text string identifying any random-effect
  summands.  Possibly blank in the case none are desired.}
  \item{\code{ziFormula}}{A text string identifying any zero-inflation
  formulae.  Possibly blank in the case none are desired.}
  }
}
\author{
WEST, Inc.
}
