% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTallBars.R
\name{makeTallBars}
\alias{makeTallBars}
\title{Create a bar chart of means with 95\% confidence intervals bounded
  below by zero.}
\usage{
makeTallBars(
  df,
  fileSave,
  title,
  xlab,
  ylab,
  dir,
  height,
  width,
  hline1 = FALSE,
  hcol = "red",
  text = 11
)
}
\arguments{
\item{df}{A \code{data.frame} containing columns for \code{FlowerType},
\code{Variety}, and \code{b0}.  Typically the output from function
\code{getPoissonMeans}.}

\item{fileSave}{A character string used to name the plot file.}

\item{title}{A character string used to entitle the plot.}

\item{xlab}{A character string used to label the \eqn{x}-axis.}

\item{ylab}{A character string used to label the \eqn{y}-axis.}

\item{dir}{A character string used to identify the directory into which
output is to be saved.  The directory should have a trailing slash
"\code{/}."}

\item{height}{A number indicating the height of the resulting \code{png}, in
inches.}

\item{width}{A number indicating the width of the resulting \code{png}, in
inches.}

\item{hline1}{A logical indicating if a vertical line should be drawn where
the \eqn{x}-axis equals one.  Default is \code{FALSE}.}

\item{hcol}{A character string used to identify the uniform color of all
bars, when only one color is needed, typically because there is one
\code{FlowerType}.  Default is \code{"red"}.}

\item{text}{The size of the output font.  Larger means bigger.  Default is
\code{11}.}
}
\value{
A \code{ggplot2}-styled bar chart, with bars oriented horizontally.
}
\description{
Create a bar chart of means with 95\% confidence intervals
  bounded below by zero.
}
\details{
This function creates a \code{png} via \code{ggplot2}.
}
\seealso{
\code{ggplot2::ggplot} and friends, \code{forcats}.
}
\author{
Jason Mitchell, based on original work by Zoe Gustafson.
}
