% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollinatorAttractiveness.R
\docType{package}
\name{pollinatorAttractiveness-package}
\alias{pollinatorAttractiveness}
\alias{pollinatorAttractiveness-package}
\title{pollinatorAttractiveness: Pollinator Attractiveness Analysis}
\description{
Package build to analyze pollinator attractiveness data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lar-git.west-inc.com/michigan-state-university/pollinatorAttractiveness}
  \item Report bugs at \url{https://lar-git.west-inc.com/michigan-state-university/pollinatorAttractiveness/issues}
}

}
\author{
\strong{Maintainer}: Jason Mitchell \email{jmitchell@west-inc.com}

}
