% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1plot.R
\name{table1plot}
\alias{table1plot}
\title{Create a Table-1-like statistical summary of experimental results.}
\usage{
table1plot(df, factor, anc)
}
\arguments{
\item{df}{A \code{data.frame} containing columns for \code{PollinatorVisits},
\code{Cultivar}, \code{SamplingPeriods}, and \code{FlowerType}.  Typically,
\code{Cultivar} is nested within \code{FlowerType}.}

\item{factor}{The variable of factor-type for whose levels form the
subsequent rows in the final table. This is typically \code{"Cultivar"}.}

\item{anc}{The grouping variable into which the \code{factor} variable
partitions.  This is typically \code{"FlowerType"}.}
}
\value{
A data frame, with each row containing summary statistics for each
  \code{factor}.
}
\description{
Create a Table-1-like statistical summary of experimental
  results.
}
\details{
Function \code{table1plot} helps to calculate a Table 1-like set of
  statistical results; i.e., simple metrics including mean, variance, and so
  on.

  Contrary to its name, \code{table1plot} doesn't actually plot anything.
}
